requirements.txt
bcrypt==4.0.1
certifi==2025.8.3
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
huggingface-hub==0.35.0
joblib==1.5.2
numpy==2.3.3
pandas==2.2.3
passlib==1.7.4
pydantic==2.10.6
pydantic-extra-types==2.10.5
pydantic-settings==2.11.0
pydantic_core==2.27.2
pytest==7.4.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.4.0
python-multipart==0.0.6
scikit-learn==1.7.2
uvicorn==0.37.0

torch
yfinance
matplotlib
mlflow